move repositoryFactory.ts
